+++ hmac.go
@@ -1,9 +1,8 @@
 package jwt
 
 import (
+	"crypto"
 	"crypto/hmac"
-	"crypto/sha256"
-	"crypto/sha512"
 	"errors"
 	"hash"
 )
@@ -15,59 +14,50 @@ var (
 	ErrHMACVerification = errors.New("jwt: HMAC verification failed")
 )
 
-type hmacsha struct {
+type HMAC struct {
 	key  []byte
-	hash func() hash.Hash
-	alg  string
+	hash crypto.Hash
+	pool *pool
 }
 
-// NewHS256 creates a signing method using HMAC and SHA-256.
-func NewHS256(key string) Signer {
-	return &hmacsha{key: []byte(key), hash: sha256.New, alg: MethodHS256}
-}
-
-// NewHS384 creates a signing method using HMAC and SHA-384.
-func NewHS384(key string) Signer {
-	return &hmacsha{key: []byte(key), hash: sha512.New384, alg: MethodHS384}
-}
-
-// NewHS512 creates a signing method using HMAC and SHA-512.
-func NewHS512(key string) Signer {
-	return &hmacsha{key: []byte(key), hash: sha512.New, alg: MethodHS512}
-}
-
-func (h *hmacsha) Sign(payload []byte) ([]byte, error) {
-	sig, err := h.sign(payload)
-	if err != nil {
-		return nil, err
+func NewHMAC(sha Hash, key []byte) *HMAC {
+	hh := sha.hash()
+	return &HMAC{
+		key:  key,
+		hash: hh,
+		pool: newPool(func() hash.Hash { return hmac.New(hh.New, key) }),
 	}
-	return build(h, payload, sig), nil
 }
 
-func (h *hmacsha) Verify(payload, sig []byte) (err error) {
-	if sig, err = decodeToBytes(sig); err != nil {
-		return err
-	}
-	return h.verify(payload, sig)
-}
-
-func (h *hmacsha) String() string {
-	return h.alg
-}
-
-func (h *hmacsha) sign(payload []byte) ([]byte, error) {
+func (h *HMAC) Sign(payload []byte) ([]byte, error) {
 	if string(h.key) == "" {
 		return nil, ErrNoHMACKey
 	}
-	hh := hmac.New(h.hash, h.key)
-	if _, err := hh.Write(payload); err != nil {
-		return nil, err
-	}
-	return hh.Sum(nil), nil
+	return h.pool.sign(payload)
 }
 
-func (h *hmacsha) verify(payload, sig []byte) error {
-	sig2, err := h.sign(payload)
+func (h *HMAC) Size() int {
+	return h.hash.Size()
+}
+
+func (h *HMAC) String() string {
+	switch h.hash {
+	case crypto.SHA256:
+		return MethodHS256
+	case crypto.SHA384:
+		return MethodHS384
+	case crypto.SHA512:
+		return MethodHS512
+	default:
+		return ""
+	}
+}
+
+func (h *HMAC) Verify(payload, sig []byte) (err error) {
+	if sig, err = decodeToBytes(sig); err != nil {
+		return err
+	}
+	sig2, err := h.Sign(payload)
 	if err != nil {
 		return err
 	}