+++ hmac_test.go
@@ -3,23 +3,23 @@ package jwt_test
 import (
 	"testing"
 
-	. "github.com/gbrlsnchs/jwt/v2"
+	. "github.com/gbrlsnchs/jwt/v3"
 )
 
 func TestHMAC(t *testing.T) {
 	testCases := []testCase{
-		{NewHS256(""), NewHS256(""), nil, ErrNoHMACKey, nil, nil, ErrNoHMACKey},
-		{NewHS256("secret"), NewHS256("secret"), nil, nil, nil, nil, nil},
-		{NewHS256("secret"), NewHS256("not_secret"), nil, nil, nil, nil, ErrHMACVerification},
-		{NewHS256("not_secret"), NewHS256("secret"), nil, nil, nil, nil, ErrHMACVerification},
-		{NewHS384(""), NewHS384(""), nil, ErrNoHMACKey, nil, nil, ErrNoHMACKey},
-		{NewHS384("secret"), NewHS384("secret"), nil, nil, nil, nil, nil},
-		{NewHS384("secret"), NewHS384("not_secret"), nil, nil, nil, nil, ErrHMACVerification},
-		{NewHS384("not_secret"), NewHS384("secret"), nil, nil, nil, nil, ErrHMACVerification},
-		{NewHS512(""), NewHS512(""), nil, ErrNoHMACKey, nil, nil, ErrNoHMACKey},
-		{NewHS512("secret"), NewHS512("secret"), nil, nil, nil, nil, nil},
-		{NewHS512("secret"), NewHS512("not_secret"), nil, nil, nil, nil, ErrHMACVerification},
-		{NewHS512("not_secret"), NewHS512("secret"), nil, nil, nil, nil, ErrHMACVerification},
+		{NewHMAC(SHA256, []byte("")), NewHMAC(SHA256, []byte("")), nil, ErrNoHMACKey, nil, nil, ErrNoHMACKey},
+		{NewHMAC(SHA256, []byte("secret")), NewHMAC(SHA256, []byte("secret")), nil, nil, nil, nil, nil},
+		{NewHMAC(SHA256, []byte("secret")), NewHMAC(SHA256, []byte("not_secret")), nil, nil, nil, nil, ErrHMACVerification},
+		{NewHMAC(SHA256, []byte("not_secret")), NewHMAC(SHA256, []byte("secret")), nil, nil, nil, nil, ErrHMACVerification},
+		{NewHMAC(SHA384, []byte("")), NewHMAC(SHA384, []byte("")), nil, ErrNoHMACKey, nil, nil, ErrNoHMACKey},
+		{NewHMAC(SHA384, []byte("secret")), NewHMAC(SHA384, []byte("secret")), nil, nil, nil, nil, nil},
+		{NewHMAC(SHA384, []byte("secret")), NewHMAC(SHA384, []byte("not_secret")), nil, nil, nil, nil, ErrHMACVerification},
+		{NewHMAC(SHA384, []byte("not_secret")), NewHMAC(SHA384, []byte("secret")), nil, nil, nil, nil, ErrHMACVerification},
+		{NewHMAC(SHA512, []byte("")), NewHMAC(SHA512, []byte("")), nil, ErrNoHMACKey, nil, nil, ErrNoHMACKey},
+		{NewHMAC(SHA512, []byte("secret")), NewHMAC(SHA512, []byte("secret")), nil, nil, nil, nil, nil},
+		{NewHMAC(SHA512, []byte("secret")), NewHMAC(SHA512, []byte("not_secret")), nil, nil, nil, nil, ErrHMACVerification},
+		{NewHMAC(SHA512, []byte("not_secret")), NewHMAC(SHA512, []byte("secret")), nil, nil, nil, nil, ErrHMACVerification},
 	}
 	testJWT(t, testCases)
 }