+++ methods.go
@@ -19,6 +19,14 @@ const (
 	MethodES384 = "ES384"
 	// MethodES512 is the method name for ECDSA and SHA-512.
 	MethodES512 = "ES512"
+	// MethodPS256 is the method name for RSA-PSS and SHA-256.
+	MethodPS256 = "PS256"
+	// MethodPS384 is the method name for RSA-PSS and SHA-384.
+	MethodPS384 = "PS384"
+	// MethodPS512 is the method name for RSA-PSS and SHA-512.
+	MethodPS512 = "PS512"
+	// MethodEd25519 is the method name for EdDSA using Ed25519 and SHA-512.
+	MethodEd25519 = "Ed25519"
 	// MethodNone is the method name for an unsecured JWT.
 	MethodNone = "none"
 )