+++ rsa_test.go
@@ -5,7 +5,7 @@ import (
 	"crypto/rsa"
 	"testing"
 
-	. "github.com/gbrlsnchs/jwt/v2"
+	. "github.com/gbrlsnchs/jwt/v3"
 )
 
 func TestRSA(t *testing.T) {
@@ -18,18 +18,30 @@ func TestRSA(t *testing.T) {
 		t.Fatal(err)
 	}
 	testCases := []testCase{
-		{NewRS256(priv, nil), NewRS256(nil, &priv.PublicKey), nil, nil, nil, nil, nil},
-		{NewRS256(priv, nil), NewRS256(nil, &priv2.PublicKey), nil, nil, nil, nil, rsa.ErrVerification},
-		{NewRS256(nil, nil), NewRS256(nil, nil), nil, ErrRSANilPrivKey, nil, nil, nil},
-		{NewRS256(priv, nil), NewRS256(nil, nil), nil, nil, nil, nil, ErrRSANilPubKey},
-		{NewRS384(priv, nil), NewRS384(nil, &priv.PublicKey), nil, nil, nil, nil, nil},
-		{NewRS384(priv, nil), NewRS384(nil, &priv2.PublicKey), nil, nil, nil, nil, rsa.ErrVerification},
-		{NewRS384(nil, nil), NewRS384(nil, nil), nil, ErrRSANilPrivKey, nil, nil, nil},
-		{NewRS384(priv, nil), NewRS384(nil, nil), nil, nil, nil, nil, ErrRSANilPubKey},
-		{NewRS512(priv, nil), NewRS512(nil, &priv.PublicKey), nil, nil, nil, nil, nil},
-		{NewRS512(priv, nil), NewRS512(nil, &priv2.PublicKey), nil, nil, nil, nil, rsa.ErrVerification},
-		{NewRS512(nil, nil), NewRS512(nil, nil), nil, ErrRSANilPrivKey, nil, nil, nil},
-		{NewRS512(priv, nil), NewRS512(nil, nil), nil, nil, nil, nil, ErrRSANilPubKey},
+		{NewRSA(SHA256, priv, nil), NewRSA(SHA256, nil, &priv.PublicKey), nil, nil, nil, nil, nil},
+		{NewRSA(SHA256, priv, nil).WithPSS(), NewRSA(SHA256, nil, &priv.PublicKey).WithPSS(), nil, nil, nil, nil, nil},
+		{NewRSA(SHA256, priv, nil), NewRSA(SHA256, nil, &priv2.PublicKey), nil, nil, nil, nil, rsa.ErrVerification},
+		{NewRSA(SHA256, priv, nil).WithPSS(), NewRSA(SHA256, nil, &priv2.PublicKey).WithPSS(), nil, nil, nil, nil, rsa.ErrVerification},
+		{NewRSA(SHA256, nil, nil), NewRSA(SHA256, nil, nil), nil, ErrRSANilPrivKey, nil, nil, nil},
+		{NewRSA(SHA256, nil, nil).WithPSS(), NewRSA(SHA256, nil, nil).WithPSS(), nil, ErrRSANilPrivKey, nil, nil, nil},
+		{NewRSA(SHA256, priv, nil), NewRSA(SHA256, nil, nil), nil, nil, nil, nil, ErrRSANilPubKey},
+		{NewRSA(SHA256, priv, nil).WithPSS(), NewRSA(SHA256, nil, nil).WithPSS(), nil, nil, nil, nil, ErrRSANilPubKey},
+		{NewRSA(SHA384, priv, nil), NewRSA(SHA384, nil, &priv.PublicKey), nil, nil, nil, nil, nil},
+		{NewRSA(SHA384, priv, nil).WithPSS(), NewRSA(SHA384, nil, &priv.PublicKey).WithPSS(), nil, nil, nil, nil, nil},
+		{NewRSA(SHA384, priv, nil), NewRSA(SHA384, nil, &priv2.PublicKey), nil, nil, nil, nil, rsa.ErrVerification},
+		{NewRSA(SHA384, priv, nil).WithPSS(), NewRSA(SHA384, nil, &priv2.PublicKey).WithPSS(), nil, nil, nil, nil, rsa.ErrVerification},
+		{NewRSA(SHA384, nil, nil), NewRSA(SHA384, nil, nil), nil, ErrRSANilPrivKey, nil, nil, nil},
+		{NewRSA(SHA384, nil, nil).WithPSS(), NewRSA(SHA384, nil, nil).WithPSS(), nil, ErrRSANilPrivKey, nil, nil, nil},
+		{NewRSA(SHA384, priv, nil), NewRSA(SHA384, nil, nil), nil, nil, nil, nil, ErrRSANilPubKey},
+		{NewRSA(SHA384, priv, nil).WithPSS(), NewRSA(SHA384, nil, nil).WithPSS(), nil, nil, nil, nil, ErrRSANilPubKey},
+		{NewRSA(SHA512, priv, nil), NewRSA(SHA512, nil, &priv.PublicKey), nil, nil, nil, nil, nil},
+		{NewRSA(SHA512, priv, nil).WithPSS(), NewRSA(SHA512, nil, &priv.PublicKey).WithPSS(), nil, nil, nil, nil, nil},
+		{NewRSA(SHA512, priv, nil), NewRSA(SHA512, nil, &priv2.PublicKey), nil, nil, nil, nil, rsa.ErrVerification},
+		{NewRSA(SHA512, priv, nil).WithPSS(), NewRSA(SHA512, nil, &priv2.PublicKey).WithPSS(), nil, nil, nil, nil, rsa.ErrVerification},
+		{NewRSA(SHA512, nil, nil), NewRSA(SHA512, nil, nil), nil, ErrRSANilPrivKey, nil, nil, nil},
+		{NewRSA(SHA512, nil, nil).WithPSS(), NewRSA(SHA512, nil, nil).WithPSS(), nil, ErrRSANilPrivKey, nil, nil, nil},
+		{NewRSA(SHA512, priv, nil), NewRSA(SHA512, nil, nil), nil, nil, nil, nil, ErrRSANilPubKey},
+		{NewRSA(SHA512, priv, nil).WithPSS(), NewRSA(SHA512, nil, nil).WithPSS(), nil, nil, nil, nil, ErrRSANilPubKey},
 	}
 	testJWT(t, testCases)
 }