+++ token.go
@@ -1,0 +1,5 @@
+package jwt
+
+type Token interface {
+	HeaderAddr() *Header // TODO(gbrlsnchs): use contracts to capture Header field
+}